HHH-11478 - General package restructuring HHH-10873 - Type systems consolidation + org.hibernate.type refactoring

+++ hibernate-core/src/main/java/org/hibernate/type/descriptor/java/internal/LongJavaDescriptor.java
@@ -34,7 +34,7 @@ public class LongJavaDescriptor extends AbstractNumericJavaDescriptor<Long> impl
 
 	@Override
 	public SqlTypeDescriptor getJdbcRecommendedSqlType(JdbcRecommendedSqlTypeMappingContext context) {
-		return context.getTypeConfiguration().getSqlTypeDescriptorRegistry().getDescriptor( Types.INTEGER );
+		return context.getTypeConfiguration().getSqlTypeDescriptorRegistry().getDescriptor( Types.BIGINT );
 	}
 
 	@Override
